requirements.txt
Flask==2.3.0
Flask-Login==0.6.2
Flask-SocketIO==5.3.0
Flask-SQLAlchemy==3.0.5
python-socketio==5.9.0
Werkzeug==2.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0